// types.ts
export enum AuthMode {
    LOGIN,
    SIGNUP_FULLNAME,
    SIGNUP_USERNAME,
    SIGNUP_EMAIL,
    SIGNUP_PASSWORD,
    SIGNUP_CONFIRM_PASSWORD,
    SIGNUP_VERIFY_CODE,
}

export enum VoiceState {
    IDLE,
    LISTENING,
    PROCESSING
}